+++ sqloquent/relations.py
@@ -1152,8 +1152,11 @@ class Contains(HasMany):
         if self.primary and self.foreign_id_column in self.primary.data:
             secondary_ids = self.primary.data[self.foreign_id_column]
             secondary_ids = secondary_ids.split(',') if secondary_ids else []
-            self._secondary = self.secondary_class.query().is_in(
-                self.secondary_class.id_column, secondary_ids).get()
+            if secondary_ids:
+                self._secondary = self.secondary_class.query().is_in(
+                    self.secondary_class.id_column, secondary_ids).get()
+            else:
+                self._secondary = tuple()
             return self
 
         if self.secondary and all([
@@ -1173,8 +1176,10 @@ class Contains(HasMany):
     def query(self) -> QueryBuilderProtocol|None:
         """Creates the base query for the underlying relation."""
         if self.primary and self.foreign_id_column in self.primary.data:
-            secondary_ids = self.primary.data[self.foreign_id_column] or ''
-            secondary_ids = secondary_ids.split(',')
+            secondary_ids = self.primary.data[self.foreign_id_column]
+            secondary_ids = secondary_ids.split(',') if secondary_ids else []
+            if not secondary_ids:
+                return self.secondary_class.query().equal('1', '2')
             return self.secondary_class.query().is_in(
                 self.secondary_class.id_column, secondary_ids
             )
@@ -1183,6 +1188,8 @@ class Contains(HasMany):
                 s.data[self.secondary_class.id_column]
                 for s in self.secondary
             ]
+            if not secondary_ids:
+                return self.secondary_class.query().equal('1', '2')
             return self.secondary_class.query().is_in(
                 self.secondary_class.id_column, secondary_ids
             )