+++ tests/test_relations.py
@@ -1471,6 +1471,18 @@ class TestRelations(unittest.TestCase):
         child = self.DAGItem.insert({'details': '321'})
         assert len(child.parents) == 0
 
+        try:
+            child.parents().reload()
+            assert True, 'contains().reload() should not error on empty column'
+        except BaseException as e:
+            assert False, 'contains().reload() should not error on empty column'
+
+        try:
+            child.parents().query()
+            assert True, 'contains().query() should not error on empty column'
+        except BaseException as e:
+            assert False, 'contains().query() should not error on empty column'
+
     # Within tests
     def test_Within_extends_Relation(self):
         assert issubclass(relations.Within, relations.Relation)